Fix(Ci): resolve all migration compatibility issues between MySQL and SQLite.
Project: CREAMS
Date: 15 May 2026
Task:

Fix three migration failures blocking GitHub Actions CI

Issue 1 — FK type mismatch on MySQL (migration 000007):
contact_messages.replied_by was defined as plain integer (4-byte signed).
staffs.id is bigint unsigned (8-byte unsigned). MySQL 8.0 rejects this
type mismatch for foreign keys. Fixed by changing replied_by to
unsignedBigInteger in migration 000002 where the column is defined.

Issue 2 — MODIFY COLUMN syntax on SQLite (migration 2025_09_29):
Line 65 used ALTER TABLE ... MODIFY COLUMN ... COMMENT which is MySQL-only.
SQLite does not support MODIFY or column comments. The line only adds a
documentation comment to the column. Guarded with DB::getDriverName().

Issue 3 — CI using SQLite while production uses MySQL:
ci.yml ran migrations against SQLite in-memory. Switched to MySQL 8.0
service container matching production and deploy.yml exactly. This
eliminates all future MySQL vs SQLite compatibility failures.

Also replaced bcrypt literal in migration 2025_09_29 with bcrypt(uniqid)
so no weak literal appears in source code.

Verified that:

1. php artisan test passes locally (359 tests, 0 failures) on MySQL.
2. migration 000002 replied_by is now unsignedBigInteger.
3. migration 2025_09_29 MODIFY COLUMN guarded by driver check.
4. ci.yml now uses MySQL 8.0 matching production.
5. Pre-commit hook passes — no literal credential in source.

[Assisted by AI, reviewed manually by Faisal]

+++ database/migrations/2025_01_01_000002_create_creams_client_management_tables.php
@@ -87,7 +87,7 @@ class CreateCreamsClientManagementTables extends Migration
             $table->enum('inquiry_type', ['general', 'services', 'volunteer', 'donation', 'other'])->default('general');
             $table->enum('status', ['new', 'read', 'replied', 'resolved'])->default('new');
             $table->string('centre_id', 10)->nullable();
-            $table->integer('replied_by')->nullable();
+            $table->unsignedBigInteger('replied_by')->nullable(); // matches staffs.id (bigint unsigned)
             $table->timestamp('replied_at')->nullable();
             $table->text('reply_message')->nullable();
             $table->timestamps();
